Create superclass BaseListVC for LibraryVC and PlaylistVC
