+++ BookPlayer/Extensions/Extensions.swift
@@ -14,6 +14,7 @@ extension Notification.Name {
         public static let requestReview = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.requestreview")
         public static let updatePercentage = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.percentage")
         public static let updateChapter = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.chapter")
+        public static let loadingBook = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.loading")
         public static let errorLoadingBook = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.error")
         public static let bookReady = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.ready")
         public static let bookPlayed = Notification.Name(rawValue: "com.tortugapower.audiobookplayer.book.play")
